improve explorer menu; add slideshow button to browser; other minor changes;

+++ Helpers/RegistryHelpers.cs
@@ -46,37 +46,57 @@ namespace ZipImageViewer
         #endregion
 
         #region Windows Explorer Context Menu
-        public static bool CheckExplorerMenuItem(string itemKey, params string[] clsSubDirs) {
+        public static bool CheckExplorerMenuItem(params string[] clsSubDirs) {
             return clsSubDirs.All(clsDir => {
-                using (var itmKey = Registry.CurrentUser.OpenSubKey($@"Software\Classes\{clsDir}\shell\{itemKey}")) {
+                using (var itmKey = Registry.CurrentUser.OpenSubKey($@"Software\Classes\{clsDir}\shell\{nameof(ZipImageViewer)}")) {
                     return itmKey != null;
                 }
             });
         }
 
-        public static void SetExplorerMenuItem(string itemKey, string itemText, string command, params string[] clsSubDirs) {
+        /// <summary>
+        /// Enable menu for certain types of files.
+        /// </summary>
+        /// <param name="clsSubDirs">The types (keys in Software\Classes) to add menu for.</param>
+        public static void SetExplorerMenuItem(params string[] clsSubDirs) {
+            // create submenu
+            using (var itmKey = Registry.CurrentUser.CreateSubKey($@"Software\Classes\{nameof(ZipImageViewer)}\shell\OpenWith")) {
+                itmKey.SetValue(@"MUIVerb", Helpers.GetRes(@"ttl_OpenWithZIV"), RegistryValueKind.String);
+                using (var cmdKey = itmKey.CreateSubKey(@"command")) {
+                    cmdKey.SetValue(null, $@"""{App.ExePath}"" ""%1""", RegistryValueKind.String);
+                }
+            }
+            using (var itmKey = Registry.CurrentUser.CreateSubKey($@"Software\Classes\{nameof(ZipImageViewer)}\shell\PlaySlideshow")) {
+                itmKey.SetValue(@"MUIVerb", Helpers.GetRes(@"ttl_PlaySlideshowWithZIV"), RegistryValueKind.String);
+                using (var cmdKey = itmKey.CreateSubKey(@"command")) {
+                    cmdKey.SetValue(null, $@"""{App.ExePath}"" -slideshow ""%1""", RegistryValueKind.String);
+                }
+            }
+
             foreach (var clsDir in clsSubDirs) {
-                using (var itmKey = Registry.CurrentUser.CreateSubKey($@"Software\Classes\{clsDir}\shell\{itemKey}")) {
-                    itmKey.SetValue(null, itemText, RegistryValueKind.String);
-                    using (var cmdKey = itmKey.CreateSubKey(@"command")) {
-                        cmdKey.SetValue(null, command, RegistryValueKind.String);
-                    }
+                // create ref to submenu
+                using (var zivKey = Registry.CurrentUser.CreateSubKey($@"Software\Classes\{clsDir}\shell\{nameof(ZipImageViewer)}")) {
+                    zivKey.SetValue(@"Icon", $@"""{App.ExePath}""", RegistryValueKind.String);
+                    zivKey.SetValue(@"ExtendedSubCommandsKey", nameof(ZipImageViewer), RegistryValueKind.String);
                 }
             }
         }
 
-        public static void ClearExplorerMenuItem(string itemKey, params string[] clsSubDirs) {
+        public static void ClearExplorerMenuItem(params string[] clsSubDirs) {
+            // delete ref to submenu
             foreach (var clsDir in clsSubDirs) {
                 using (var dirKey = Registry.CurrentUser.OpenSubKey($@"Software\Classes\{clsDir}", true)) {
                     if (dirKey == null) continue;
-                    dirKey.DeleteSubKeyTree($@"shell\{itemKey}", false);
+                    dirKey.DeleteSubKeyTree($@"shell\{nameof(ZipImageViewer)}", false);
                     //shell key might be created by this program. Delete it when nothing is underneath.
                     using (var shlKey = dirKey.OpenSubKey(@"shell", true)) {
-                        if (shlKey != null && shlKey.SubKeyCount == 0 && shlKey.ValueCount == 0)
-                            dirKey.DeleteSubKeyTree("shell");
+                        if (shlKey != null && shlKey.SubKeyCount == 0 && shlKey.ValueCount == 0) dirKey.DeleteSubKeyTree("shell");
                     }
                 }
             }
+
+            // delete submenu
+            Registry.CurrentUser.DeleteSubKeyTree($@"Software\Classes\{nameof(ZipImageViewer)}", false);
         }
         #endregion
     }
